tone equalizer : prototype [WIP]
introduce a new "tone equalizer " module, intended to allow selective exposure adjustements in linear RGB (supposed to be color-safe).

This is very similar to Lightroom, with blacks/shadows/midtones/highlights/whites cursors.

The luminance channels are blended with gaussian masks and the modifications are applied on a gaussian pyramid to preserve the local contrast (luminance modified in low frequency layers).

For now, nothing is optimized (naive C implementation). This is just a proof of concept.