lighttable: make lighttable faster on High DPI screen.
The lighttable was quite slow due to lot of expose events being
sent (one expose event for each mouse move). And in this case a
full redraw of the lighttable plus the thumbs was done. This
resulted in darktable using all the CPU.

Many mouse_moved events are now filtered out (some mesurement shows that
only 15% of the mouse_move are actually handled).

With the patch the lighttable should be usable on 4k and 5k screens.